+++ src/routes/scholar/[id]/+page.svelte
@@ -6,6 +6,21 @@
 
 	let { data } = $props();
 
+	let { commitments, venues } = $derived(data);
+
+	let volunteering = $derived(
+		commitments
+			?.map((c) => {
+				return {
+					name: c.roles?.name,
+					venue: venues?.find((v) => v.id === c.roles?.venueid)?.title
+				};
+			})
+			.filter(
+				(v): v is { name: string; venue: string } => v.name !== undefined && v.venue !== undefined
+			) ?? []
+	);
+
 	let db = getDB();
 
 	/** Register a scholar state. */
@@ -14,7 +29,7 @@
 
 {#if state}
 	<Page title={state.getName() ?? 'anonymous'} subtitle="Scholar">
-		<ScholarView scholar={state} />
+		<ScholarView scholar={state} commitments={volunteering} />
 	</Page>
 {:else}
 	<h1>Oops.</h1>