+++ CHANGELOG.md
@@ -7,7 +7,9 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 
 ### Added
 
-- Added ability to volunteer for a role.
+- Added ability to volunteer for a role and set expertise.
+- Added ability to stop volunteering for a role.
+- Added list of volunteer roles to profile.
 
 ### Fixed